Refactor seeker endpoints
- Now each endpoint and method are in their own folder and file structure.
- Switched to using an in-memory database for the moment to make testing easier.
