+++ siftseek/endpoints/seeker/__init__.py
@@ -1,0 +1,9 @@
+from flask import Blueprint
+
+seeker = Blueprint("seeker", __name__)
+
+from .profile import get as profile_get
+from .profile import delete as profile_delete
+from .profile import patch as profile_patch
+from .profile import post as profile_post
+from .profile import put as profile_put
