--- siftseek/routes/seeker/__init__.py
@@ -1,5 +1,0 @@
-from flask import Blueprint
-
-seeker = Blueprint("seeker", __name__)
-
-from . import seeker_views
